server/requirements/common.txt
django==2.1.4
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.1.5
django-cors-headers==3.0.2